+++ OpenDentBusiness/TableTypes/InsPlan.cs
@@ -56,7 +56,7 @@ namespace OpenDentBusiness{
 		///<summary>FK to insfilingcode.InsFilingCodeNum.  Used for e-claims.  Also used for some complex reports in public health.  The e-claim usage might become obsolete when PlanID implemented by HIPAA.  Can be 0 to indicate none.  Then 'CI' will go out on claims.</summary>
 		public long FilingCode;
 		///<summary>Canadian e-claim field. D11 and E07.  Mandatory for Dentaide.  Value must be greater than 0.  Not used for all others.  2 digit.</summary>
-		public long DentaideCardSequence;
+		public int DentaideCardSequence;
 		///<summary>If checked, the units Qty will show the base units assigned to a procedure on the claim form.</summary>
 		public bool ShowBaseUnits;
 		///<summary>Apply deductible before percentage when calculating estimates.</summary>
@@ -66,9 +66,9 @@ namespace OpenDentBusiness{
 		///<summary>Set to true to hide it from the pick list and from the main list.</summary>
 		public bool IsHidden;
 		///<summary>The month, 1 through 12 when the insurance plan renews.  It will renew on the first of the month.  To indicate calendar year, set renew month to 0.</summary>
-		public long MonthRenew;
+		public int MonthRenew;
 		///<summary>This is not a database column.  It is just used to display the number of plans with the same info.</summary>
-		public long NumberPlans;
+		public int NumberPlans;
 		///<summary>FK to insfilingcodesubtype.insfilingcodesubtypenum</summary>
 		public long FilingCodeSubtype;
