+++ OpenDentBusiness/TableTypes/LabTurnaround.cs
@@ -13,9 +13,9 @@ namespace OpenDentBusiness{
 		///<summary>The description of the service that the lab is performing.</summary>
 		public string Description;
 		///<summary>The number of days that the lab publishes as the turnaround time for the service.</summary>
-		public long DaysPublished;
+		public int DaysPublished;
 		///<summary>The actual number of days.  Might be longer than DaysPublished due to travel time.  This is what the actual calculations will be done on.</summary>
-		public long DaysActual;
+		public int DaysActual;
 
 		public LabTurnaround Copy(){
 			LabTurnaround l=new LabTurnaround();
