+++ OpenDentBusiness/TableTypes/PlannedAppt.cs
@@ -43,10 +43,10 @@ namespace OpenDentBusiness{
 		}
 
 		[DataField("ItemOrder")]
-		private long itemOrder;
+		private int itemOrder;
 		private bool itemOrderChanged;
 		///<summary>One-indexed order of item in group of planned appts.</summary>
-		public long ItemOrder{
+		public int ItemOrder{
 			get{return itemOrder;}
 			set{if(itemOrder!=value){itemOrder=value;MarkDirty();itemOrderChanged=true;}}
 		}
