+++ OpenDentBusiness/TableTypes/ProcTP.cs
@@ -14,7 +14,7 @@ namespace OpenDentBusiness{
 		///<summary>FK to procedurelog.ProcNum.  It is very common for the referenced procedure to be missing.  This procNum is only here to compare and test the existence of the referenced procedure.  If present, it will check to see whether the procedure is still status TP.</summary>
 		public long ProcNumOrig;
 		///<summary>The order of this proc within its tp.  This is set when the tp is first created and can't be changed.  Drastically simplifies loading the tp.</summary>
-		public long ItemOrder;
+		public int ItemOrder;
 		///<summary>FK to definition.DefNum which contains the text of the priority.</summary>
 		public long Priority;
 		///<summary>A simple string displaying the tooth number.  If international tooth numbers are used, then this will be in international format already.</summary>
