+++ OpenDentBusiness/TableTypes/SigElementDef.cs
@@ -9,7 +9,7 @@ namespace OpenDentBusiness{
 		///<summary>Primary key.</summary>
 		public long SigElementDefNum;
 		///<summary>If this element should cause a button to light up, this would be the row.  0 means none.</summary>
-		public long LightRow;
+		public int LightRow;
 		///<summary>If a light row is set, this is the color it will turn when triggered.  Ack sets it back to white.  Note that color and row can be in two separate elements of the same signal.</summary>
 		public Color LightColor;
 		///<summary>Enum:SignalElementType  0=User,1=Extra,2=Message.</summary>
@@ -19,7 +19,7 @@ namespace OpenDentBusiness{
 		///<summary>The sound to play for this element.  Wav file stored in the database in string format until "played".  If empty string, then no sound.</summary>
 		public string Sound;
 		///<summary>The order of this element within the list of the same type.</summary>
-		public long ItemOrder;
+		public int ItemOrder;
 
 		///<summary></summary>
 		public SigElementDef Copy() {
