+++ OpenDentBusiness/TableTypes/Userod.cs
@@ -24,14 +24,14 @@ namespace OpenDentBusiness{
 		///<summary>FK to tasklist.TaskListNum.  0 if no inbox setup yet.  It is assumed that the TaskList is in the main trunk, but this is not strictly enforced.  User can't delete an attached TaskList, but they could move it.</summary>
 		public long TaskListInBox;
 		/// <summary> Defaults to 3 (regular user) unless specified. Helps populates the Anesthetist, Surgeon, Assistant and Circulator dropdowns properly on FormAnestheticRecord/// </summary>
-		public long AnesthProvType;
+		public int AnesthProvType;
 
 
 		public Userod(){
 
 		}
 
-		public Userod(long userNum, string userName, string password, long userGroupNum, long employeeNum, long clinicNum, long provNum, bool isHidden, long anesthProvType)
+		public Userod(long userNum, string userName, string password, long userGroupNum, long employeeNum, long clinicNum, long provNum, bool isHidden, int anesthProvType)
 		{ 
 			UserNum=userNum;
 			UserName=userName;
